연구 센터 스테이트 사용
