+++ app/[locale]/academics/undergraduate/course-changes/page.tsx
@@ -27,7 +27,7 @@ const TIME_SPOTS: { year: number; margin?: string; isLast?: boolean }[] = [
   { year: 2013, margin: 'ml-[5.375rem]' },
   { year: 2012, margin: 'ml-7' },
   { year: 2011, margin: 'ml-7' },
-  { year: 2010, margin: 'ml-[7rem]', isLast: true },
+  { year: 2010, margin: 'ml-7', isLast: true },
 ];
 
 const getSelectedChanges = (selectedYear: number, data: CourseChange[]) => {
